Revert "Merge pull request #60 from thaodangspace/codex/refactor-server-code-for-readability"
This reverts commit 485f603ff15e2ef3bc912b53a7f0a4a5c8399882, reversing
changes made to 7cc90611c0c5e26e8c79d4fcd86f7d0ba4ee0b45.
